Add Rhisis.Business library
