+++ .travis.yml
@@ -16,12 +16,16 @@ jobs:
         name: "Rhisis Core Feature"
       - script: dotnet build src/Rhisis.Database/ --framework netstandard2.0 --configuration Release
         name: "Rhisis Database access Layer"
+      - script: dotnet build src/Rhisis.Business/ --framework netstandard2.0 --configuation Release
+        name: "Rhisis Business Layer"
       - script: dotnet build src/Rhisis.Login/ --framework netcoreapp2.0 --configuration Release
         name: "Login Server"
       - script: dotnet build src/Rhisis.Cluster/ --framework netcoreapp2.0 --configuration Release
         name: "Cluster Server"
       - script: dotnet build src/Rhisis.World/ --framework netcoreapp2.0 --configuration Release
         name: "World Server"
+      - script: dotnet build src/Rhisis.Admin/ --framework netcoreapp2.0 --configuation Release
+        name: "Administration tool"
 
       - stage: "Tests"
         script: dotnet test test/Rhisis.Core.Test/
