+++ Rhisis.sln
@@ -41,6 +41,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Rhisis.Tools.Core", "src\to
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Rhisis.Admin", "src\Rhisis.Admin\Rhisis.Admin.csproj", "{A1C2BED4-9BDF-49D5-9E07-A6306DEF1BF7}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Rhisis.Business", "src\Rhisis.Business\Rhisis.Business.csproj", "{CE55255B-935A-4EF2-B780-84CE9E1E9EE3}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -91,6 +93,10 @@ Global
 		{A1C2BED4-9BDF-49D5-9E07-A6306DEF1BF7}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A1C2BED4-9BDF-49D5-9E07-A6306DEF1BF7}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{A1C2BED4-9BDF-49D5-9E07-A6306DEF1BF7}.Release|Any CPU.Build.0 = Release|Any CPU
+		{CE55255B-935A-4EF2-B780-84CE9E1E9EE3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{CE55255B-935A-4EF2-B780-84CE9E1E9EE3}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{CE55255B-935A-4EF2-B780-84CE9E1E9EE3}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{CE55255B-935A-4EF2-B780-84CE9E1E9EE3}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -108,6 +114,7 @@ Global
 		{8EF0988E-C4FC-4E5A-9EA4-C7E720CF24AC} = {33862335-9BF3-4985-81B2-4136B4CAB46C}
 		{BEC2AFBB-87F5-46C8-A75C-0191614B7707} = {33862335-9BF3-4985-81B2-4136B4CAB46C}
 		{A1C2BED4-9BDF-49D5-9E07-A6306DEF1BF7} = {CC42CC8A-A3BC-4913-90A0-1AF2213A870F}
+		{CE55255B-935A-4EF2-B780-84CE9E1E9EE3} = {CC42CC8A-A3BC-4913-90A0-1AF2213A870F}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {B573C179-78E3-4F7B-8CFB-289754036BBA}
