+++ src/Rhisis.Business/Rhisis.Business.csproj
@@ -1,0 +1,11 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netcoreapp2.0</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\Rhisis.Core\Rhisis.Core.csproj" />
+  </ItemGroup>
+
+</Project>
