+++ src/Rhisis.Business/Services/AuthenticationService.cs
@@ -1,0 +1,8 @@
+﻿using Rhisis.Core.Services;
+
+namespace Rhisis.Business.Services
+{
+    public class AuthenticationService : IAuthenticationService
+    {
+    }
+}
