+++ src/Rhisis.Business/Services/UserService.cs
@@ -1,0 +1,9 @@
+﻿using Rhisis.Core.Services;
+
+namespace Rhisis.Business.Services
+{
+    public class UserService : IUserService
+    {
+        
+    }
+}
