+++ src/Rhisis.Core/Services/IAuthenticationService.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Rhisis.Core.Services
+{
+    public interface IAuthenticationService
+    {
+    }
+}
