+++ src/Rhisis.Core/Services/IUserService.cs
@@ -1,0 +1,7 @@
+﻿namespace Rhisis.Core.Services
+{
+    public interface IUserService
+    {
+        
+    }
+}
